I am mainly finished with cleaning the data in problem 3.
This involved going through the raw.csv files for pol-monts, snp, and unemployment and fixing the date structure using the seperate function. After seperating we needed to convert the months from numeric to character.

For the snp the year was in two digit form thus it needed to be corrected to 4 digit by creating a new variable that would add 2000 or 1900 depending on if the value was greater than or less than 22.

Days were deselected.

For unemployment pivot_longer was used to create two variables (month and unemployment percent).

The final step is to create the new presidents variable and then after that I will need to merge all of them and then write a paragrpah describing it. I also need to do the same for problem 2.

Victory is close at hand!